[Truffle] Created EncodingConverterPrimitiveNodes.
